Fix upload bugs

+++ SE_Backend/ContestPlus/backend_code/notice.py
@@ -39,7 +39,7 @@ def apiNoticeNew(request):
                 os.makedirs(file_dir)
 
             file_name_parts = str(file.name).split('.')
-            file.name = str(new_notice.id) + '.' + file_name_parts[1]
+            file.name = str(new_notice.id) + '.' + file_name_parts[-1]
             host_prefix = 'http://127.0.0.1:8000/static/'
             new_notice.file = file_dir+file.name
             new_notice.save()
@@ -92,11 +92,11 @@ def apiNoticeModify(request):
                 os.remove(os.path.join(file_dir, old_file_name))
 
                 file_name_parts = str(file.name).split('.')
-                file.name = str(notice[0].id) + '.' + file_name_parts[1]
+                file.name = str(notice[0].id) + '.' + file_name_parts[-1]
                 host_prefix = 'http://127.0.0.1:8000/static/'
 
                 notice[0].file = file_dir+file.name
-                notice[0].save()
+
 
                 destination = open(os.path.join(file_dir, file.name), 'wb+')
                 for chunk in file.chunks():
@@ -105,6 +105,7 @@ def apiNoticeModify(request):
             else:
                 if notice[0].file:
                     os.remove(os.path.join(notice[0].file))
+        notice[0].save()
         return JsonResponse({'message': 'ok'})
     return JsonResponse({'error': 'need POST method'})
 
@@ -128,12 +129,13 @@ def apiNoticeDelete(request):
         if utype != 'sponsor' and utype != 'admin' and user.id != target_contest[0].sponsorId:
             return JsonResponse({"error": "permission denied"})
 
-        file_dir = str(settings.BASE_DIR) + "\\Files\\ContestNotice\\" + str(notice[0].contest_id) + "\\"
-        if os.path.exists(file_dir) == False:
-            os.makedirs(file_dir)
+        if notice[0].file !='':
+            file_dir = str(settings.BASE_DIR) + "\\Files\\ContestNotice\\" + str(notice[0].contest_id) + "\\"
+            if os.path.exists(file_dir) == False:
+                os.makedirs(file_dir)
 
-        old_file_name = notice[0].file.split('/')[-1]
-        os.remove(os.path.join(file_dir, old_file_name))
+            old_file_name = notice[0].file.split('/')[-1]
+            os.remove(os.path.join(file_dir, old_file_name))
 
         notice[0].delete()
